moved bounce rate tests to a new file
