+++ tests/conftest.py
@@ -4063,254 +4063,6 @@ def mock_salesforce_get_accounts(mocker: MockerFixture):
     return mocker.patch("app.salesforce_account.get_accounts", return_value=mock_crm_orgs)
 
 
-jobs_2_failures = [
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T18:09:07.737196+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "6b4a62a7-329f-4332-8106-63ecc6cf7a1c",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success_3.csv",
-        "processing_finished": "2023-04-18T18:09:10.200722+00:00",
-        "processing_started": "2023-04-18T18:09:08.041267+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T18:09:10.202238+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T18:08:29.951155+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "b5e56c24-cfd6-439b-9c95-3c671d25545f",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success_2.csv",
-        "processing_finished": "2023-04-18T18:08:31.986638+00:00",
-        "processing_started": "2023-04-18T18:08:30.961924+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T18:08:31.987927+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T14:37:30.363833+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "298b3d50-feec-47a4-a771-1f34d860f513",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success.csv",
-        "processing_finished": "2023-04-18T14:37:35.161901+00:00",
-        "processing_started": "2023-04-18T14:37:34.168804+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T14:37:35.163298+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-    {
-        "failure_rate": 10.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T14:03:07.432308+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "8f5cae07-e4dc-4a31-9c31-afc1fddf2f30",
-        "job_status": "finished",
-        "notification_count": 20,
-        "original_file_name": "bulk_send_mix_20_2_fails.csv",
-        "processing_finished": "2023-04-18T14:03:08.605917+00:00",
-        "processing_started": "2023-04-18T14:03:07.839004+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [
-            {"count": 1, "status": "temporary-failure"},
-            {"count": 18, "status": "delivered"},
-            {"count": 1, "status": "permanent-failure"},
-        ],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T14:03:08.606670+00:00",
-        "notifications_sent": 20,
-        "notifications_delivered": 18,
-        "notifications_failed": 2,
-        "notifications_requested": 20,
-        "bounce_count": 1,
-    },
-    {
-        "failure_rate": 10.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T11:15:22.842906+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "89e0f76e-777f-4d3a-aee1-bdd9277837ff",
-        "job_status": "finished",
-        "notification_count": 20,
-        "original_file_name": "bulk_send_mix_20_2_fails copy.csv",
-        "processing_finished": "2023-04-18T11:15:40.718126+00:00",
-        "processing_started": "2023-04-18T11:15:39.333558+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 18, "status": "delivered"}, {"count": 2, "status": "permanent-failure"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T11:15:40.721100+00:00",
-        "notifications_sent": 20,
-        "notifications_delivered": 18,
-        "notifications_failed": 2,
-        "notifications_requested": 20,
-        "bounce_count": 2,
-    },
-]
-
-jobs_0_failures = [
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T18:09:07.737196+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "6b4a62a7-329f-4332-8106-63ecc6cf7a1c",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success_3.csv",
-        "processing_finished": "2023-04-18T18:09:10.200722+00:00",
-        "processing_started": "2023-04-18T18:09:08.041267+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T18:09:10.202238+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T18:08:29.951155+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "b5e56c24-cfd6-439b-9c95-3c671d25545f",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success_2.csv",
-        "processing_finished": "2023-04-18T18:08:31.986638+00:00",
-        "processing_started": "2023-04-18T18:08:30.961924+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T18:08:31.987927+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-    {
-        "failure_rate": 0.0,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T14:37:30.363833+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "298b3d50-feec-47a4-a771-1f34d860f513",
-        "job_status": "finished",
-        "notification_count": 19,
-        "original_file_name": "bulk_send_19_success.csv",
-        "processing_finished": "2023-04-18T14:37:35.161901+00:00",
-        "processing_started": "2023-04-18T14:37:34.168804+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 19, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T14:37:35.163298+00:00",
-        "notifications_sent": 19,
-        "notifications_delivered": 19,
-        "notifications_failed": 0,
-        "notifications_requested": 19,
-        "bounce_count": 0,
-    },
-]
-
-jobs_1_failure = [
-    {
-        "failure_rate": 0.001,
-        "api_key": None,
-        "archived": False,
-        "created_at": "2023-04-18T18:09:07.737196+00:00",
-        "created_by": {"id": "24baebe0-fb11-4fc0-8609-1e853c31d0fe", "name": "Andrew Leith"},
-        "id": "6b4a62a7-329f-4332-8106-63ecc6cf7a1c",
-        "job_status": "finished",
-        "notification_count": 1000,
-        "original_file_name": "bulk_send_19_success_3.csv",
-        "processing_finished": "2023-04-18T18:09:10.200722+00:00",
-        "processing_started": "2023-04-18T18:09:08.041267+00:00",
-        "scheduled_for": None,
-        "sender_id": None,
-        "service": "9cfb3884-fed6-4824-8901-c7d0857cc5b4",
-        "service_name": {"name": "Bounce Rate"},
-        "statistics": [{"count": 999, "status": "delivered"}],
-        "template": "2156a57e-efd7-4531-b8f4-e7e0c64c03dc",
-        "template_version": 1,
-        "updated_at": "2023-04-18T18:09:10.202238+00:00",
-        "notifications_sent": 1000,
-        "notifications_delivered": 999,
-        "notifications_failed": 1,
-        "notifications_requested": 1000,
-        "bounce_count": 2,
-    }
-]
-
-
 def create_api_user_active(with_unique_id=False):
     return {
         "id": str(uuid4()) if with_unique_id else sample_uuid(),
